+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 - DefaultColumnConditionModel bug fixed, case insensitivity now applies to Character as well as String values.
 ### is.codion.framework.domain
 - SingleValueColumnCondition bug fixed, useLikeOperator now included in equals() and hashCode().
+- ColumnCondition.equalToIgnoreCase(Character value) and notEqualToIgnoreCase(String value) added, related refactoring.
 ### is.codion.framework.db.local
 - DefaultLocalEntityConnection, optimistic locking now only compares values which are available in the entity being updated, otherwise excluding attributes when selecting entities prevents those entities from being updated, since optimistic locking would always fail due to the missing values.
 ### is.codion.swing.common.ui